% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{fozzie_difference_join_rs}
\alias{fozzie_difference_join_rs}
\title{Internal: Difference Join via Rust}
\usage{
fozzie_difference_join_rs(
  df1,
  df2,
  by,
  how,
  max_distance,
  distance_col,
  nthread
)
}
\description{
Internal function. Performs a difference-based fuzzy join using Rust backend.
}
\keyword{internal}
