% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{fozzie_distance_join_rs}
\alias{fozzie_distance_join_rs}
\title{Internal: Distance Join via Rust}
\usage{
fozzie_distance_join_rs(
  df1,
  df2,
  by,
  method,
  how,
  max_distance,
  distance_col,
  nthread
)
}
\description{
Internal function. Performs a distance-based fuzzy join using Rust backend.
}
\keyword{internal}
