% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{fozzie_interval_join_rs}
\alias{fozzie_interval_join_rs}
\title{Internal: Interval Join via Rust}
\usage{
fozzie_interval_join_rs(
  df1,
  df2,
  by,
  how,
  overlap_type,
  maxgap,
  minoverlap,
  interval_mode,
  nthread
)
}
\description{
Internal function. Performs an interval-based fuzzy join using Rust backend.
}
\keyword{internal}
