% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{fozzie_regex_join_rs}
\alias{fozzie_regex_join_rs}
\title{Internal: Regex Join via Rust}
\usage{
fozzie_regex_join_rs(df1, df2, by, how, ignore_case, nthread)
}
\description{
Internal function. Performs a regex-based fuzzy join using Rust backend.
}
\keyword{internal}
