% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{fozzie_string_join_rs}
\alias{fozzie_string_join_rs}
\title{Internal: String Join via Rust}
\usage{
fozzie_string_join_rs(
  df1,
  df2,
  by,
  method,
  how,
  max_distance,
  distance_col,
  q,
  max_prefix,
  prefix_weight,
  nthread
)
}
\description{
Internal function. Performs a string-based fuzzy join using Rust backend.
}
\keyword{internal}
