% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fozziejoin-package.R
\docType{package}
\name{fozziejoin-package}
\alias{fozziejoin}
\alias{fozziejoin-package}
\title{fozziejoin: Performant Data Frame Joins with Inexact String Matching}
\description{
Join dataframes and tibbles by inexact criteria, such as numeric difference and string distance.
}
\author{
\strong{Maintainer}: Jon Downs \email{jon@jondowns.net}

Other contributors:
\itemize{
  \item The authors of the dependency Rust crates (see inst/AUTHORS file for details) [contributor, copyright holder]
}

}
\keyword{internal}
