% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{get_nthread_default}
\alias{get_nthread_default}
\title{Get number of threads in global thread pool}
\usage{
get_nthread_default()
}
\description{
Returns default rayon number of threads
}
\keyword{internal}
