% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fozziejoin-package.R
\docType{package}
\name{fozziejoin-package}
\alias{fozziejoin}
\alias{fozziejoin-package}
\title{fozziejoin: Utilities for Joining Dataframes with Inexact Matching}
\description{
Join data frames by inexact criteria, such as numeric difference and string distance.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/fozzieverse/fozziejoin}
  \item Report bugs at \url{https://github.com/fozzieverse/fozziejoin/issues}
}

}
\author{
\strong{Maintainer}: Jon Downs \email{jon@jondowns.net}

Other contributors:
\itemize{
  \item The authors of the dependency Rust crates (see inst/AUTHORS file for details) [contributor, copyright holder]
}

}
\keyword{internal}
